D3 Pin's current voltage level is printed to the LCD via ADC.

+++ TouchGFX/gui/include/gui/screen1_screen/Screen1Presenter.hpp
@@ -27,7 +27,7 @@ public:
 
     virtual ~Screen1Presenter() {};
 
-    void setNewValue(unsigned int value);
+    void setNewValue(uint16_t value);
 
 private:
     Screen1Presenter();
